+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Removed the deprecated `Get-VCFNsxManagerCluster` alias from `Get-VCFNsxtCluster`.
 - Removed the deprecated `Get-VCFNsxEdgeCluster` alias from `Get-VCFEdgeCluster`.
 - Removed the deprecated `New-VCFNsxEdgeCluster` alias from `New-VCFEdgeCluster`.
+- Refactors `Invoke-VCFCommand` to run commands on SDDC Manager without the need for SSH across PowerShell editions and operating systems.
 
 ## v2.4.1